Support non-blocking operation.
This commit introduces a new opaque type "shoes_connstate_t." This type
is allocated and passed to the shoes_handshake function. The user will
call shoes_handshake repeatedly until shoes_is_connected returns true.
The shoes_connstate_t type is used by shoes_handshake to implement a
state machine to allow resuming of the handshake on the next call to
shoes_handshake.

This commit also removes the need for the libpack library.